pipelines/bex-to-cv/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bex-to-cv"
version = "0.1.0"
description = "Convert SAP BW BEx Query XML to HANA Calculation View XML"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Xsodus Team"}
]

dependencies = [
    "lxml>=5.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
bex-to-cv = "bex_to_cv.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"bex_to_cv.catalog.data" = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]
